Add sanitized data from 2017 Shard travel form.
Add a spreadsheet into the info folder that has the sanized version of
the enl travel plans for the 017 Spring Global Shard event.  This
similtaniusly prevents the programers from having to make up test dummy
infomation themselves, but it also gives the programs something to
guadge the program against since they will be able to compare, what the
program says they can do, with what they actually did for the legs.